Moved the main rendering into the ExploreContainer
